requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
httpx>=0.27,<0.29
psycopg2-binary>=2.9,<3
